+++ src/messages/errors/mod/ModError.java
@@ -1,4 +1,4 @@
-package messages.errors.mod_error;
+package messages.errors.mod;
 
 import messages.errors.Error;
 import utilities.PrettyPrint;
